backend/pyproject.toml
[tool.black]
line-length = 100

[tool.poetry]
name = "XXX-sandbox"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.11.0,<4.0"
numpy = "^1.25.1"
Flask = "^2.3.2"
langchain = "v0.0.235"
streamlit = "^1.25.0"
pandas = "^2.0.3"
xds-protos = "^0.0.11"
tiktoken = "^0.4.0"
setuptools = "^67.8.0"
faiss-cpu = "^1.7.4"
strif = "^0.2.3"
click = "^8.1.6"
tqdm = "^4.65.0"
cachetools = "^5.3.1"
justext = "^3.0.0"
lxml = "^4.9.3"
watchdog = "^3.0.0"
streamlit-searchbox = "^0.1.3"
flask-sqlalchemy = "^3.0.5"
marko = "2.0.0"
transformers = "^4.31.0"
marshmallow-sqlalchemy = "^0.29.0"
flask-marshmallow = "^0.15.0"
cssselect = "^1.2.0"
marshmallow-enum = "^1.5.1"
yt-dlp = "^2023.7.6"
marisa-trie = "^0.8.0"
rapidfuzz = "^3.1.2"
pydub = "^0.25.1"
webrtcvad = "^2.0.10"
pytest = "^7.4.0"
sqlalchemy-utils = "^0.41.1"
psycopg2 = "^2.9.6"
flask-migrate = "^4.0.4"
llama-index = "0.7.12"
pydantic = "1.10.11"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.7"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
black = "^23.1.0"
flake8 = "^6.0.0"
matplotlib = "^3.6.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
